Pin charm dependencies to `2/edge`
