+++ tests/integration/test_charm_auth.py
@@ -2,9 +2,8 @@
 # See LICENSE file for licensing details.
 
 import logging
-from dataclasses import asdict, dataclass
+from dataclasses import asdict
 from pathlib import Path
-from typing import Optional
 
 import pytest
 import yaml
@@ -13,6 +12,8 @@ from helpers import (
     get_configmap_data,
     get_k8s_service_address,
     get_listener_condition,
+    istio_k8s,
+    oauth_k8s,
 )
 from pytest_operator.plugin import OpsTest
 
@@ -29,41 +30,20 @@ INGRESS_CONFIG_RELATION = "istio-ingress-config"
 FORWARD_AUTH_RELATION = "forward-auth"
 
 
-@dataclass
-class CharmDeploymentConfiguration:
-    entity_url: str  # Charm name or local path to charm
-    application_name: str
-    channel: str
-    trust: bool
-    config: Optional[dict] = None
-
-
-ISTIO_K8S = CharmDeploymentConfiguration(
-    entity_url="istio-k8s", application_name="istio-k8s", channel="latest/edge", trust=True
-)
-
-OAUTH2_K8S = CharmDeploymentConfiguration(
-    entity_url="oauth2-proxy-k8s",
-    application_name="oauth2-proxy-k8s",
-    channel="latest/edge",
-    trust=True,
-)
-
-
 @pytest.mark.abort_on_fail
 async def test_deploy_dependencies(ops_test: OpsTest):
     # Deploy Istio-k8s in a separate model and oauth2 istio-ingress-k8s in the primary model.
     await ops_test.track_model(CORE_ISTIO_MODEL)
     istio_core = ops_test.models.get(CORE_ISTIO_MODEL)
 
-    await istio_core.model.deploy(**asdict(ISTIO_K8S))
+    await istio_core.model.deploy(**asdict(istio_k8s))
     await istio_core.model.wait_for_idle(
-        [ISTIO_K8S.application_name], status="active", timeout=1000
+        [istio_k8s.application_name], status="active", timeout=1000
     )
 
-    await ops_test.model.deploy(**asdict(OAUTH2_K8S))
+    await ops_test.model.deploy(**asdict(oauth_k8s))
     await ops_test.model.wait_for_idle(
-        [OAUTH2_K8S.application_name], status="active", timeout=1000
+        [oauth_k8s.application_name], status="active", timeout=1000
     )
 
 
@@ -82,7 +62,7 @@ async def test_relations_setup(ops_test: OpsTest):
     await istio_core.model.create_offer(
         endpoint=INGRESS_CONFIG_RELATION,
         offer_name=INGRESS_CONFIG_RELATION,
-        application_name=ISTIO_K8S.application_name,
+        application_name=istio_k8s.application_name,
     )
     # For some reason when using the direct consume() method, it raises a File not found error
     # await ops_test.model.consume(f"admin/{istio_core.model.name}.{INGRESS_CONFIG_RELATION}")
@@ -92,15 +72,15 @@ async def test_relations_setup(ops_test: OpsTest):
         f"admin/{istio_core.model.name}.{INGRESS_CONFIG_RELATION}",
     )
     await ops_test.model.add_relation(
-        f"{OAUTH2_K8S.application_name}:{FORWARD_AUTH_RELATION}", APP_NAME
+        f"{oauth_k8s.application_name}:{FORWARD_AUTH_RELATION}", APP_NAME
     )
     await ops_test.model.add_relation(INGRESS_CONFIG_RELATION, APP_NAME)
 
     await ops_test.model.wait_for_idle(
-        apps=[APP_NAME, OAUTH2_K8S.application_name], status="active", timeout=1000
+        apps=[APP_NAME, oauth_k8s.application_name], status="active", timeout=1000
     )
     await istio_core.model.wait_for_idle(
-        [ISTIO_K8S.application_name], status="active", timeout=1000
+        [istio_k8s.application_name], status="active", timeout=1000
     )
 
 
@@ -120,15 +100,15 @@ async def test_oauth2_proxy_relation_break_and_recovery(ops_test: OpsTest):
 
     await ops_test.juju(
         "remove-relation",
-        f"{OAUTH2_K8S.application_name}:{FORWARD_AUTH_RELATION}",
+        f"{oauth_k8s.application_name}:{FORWARD_AUTH_RELATION}",
         APP_NAME,
     )
 
     await ops_test.model.wait_for_idle(
-        apps=[APP_NAME, OAUTH2_K8S.application_name], status="active", timeout=1000
+        apps=[APP_NAME, oauth_k8s.application_name], status="active", timeout=1000
     )
     await istio_core.model.wait_for_idle(
-        [ISTIO_K8S.application_name], status="active", timeout=1000
+        [istio_k8s.application_name], status="active", timeout=1000
     )
 
     # After breaking the relation, expect the policy to be removed and the extensionProviders cleared.
@@ -140,13 +120,13 @@ async def test_oauth2_proxy_relation_break_and_recovery(ops_test: OpsTest):
 
     # Re-establish the relation and verify the config state.
     await ops_test.model.add_relation(
-        f"{OAUTH2_K8S.application_name}:{FORWARD_AUTH_RELATION}", APP_NAME
+        f"{oauth_k8s.application_name}:{FORWARD_AUTH_RELATION}", APP_NAME
     )
     await ops_test.model.wait_for_idle(
-        apps=[APP_NAME, OAUTH2_K8S.application_name], status="active", timeout=1000
+        apps=[APP_NAME, oauth_k8s.application_name], status="active", timeout=1000
     )
     await istio_core.model.wait_for_idle(
-        [ISTIO_K8S.application_name], status="active", timeout=1000
+        [istio_k8s.application_name], status="active", timeout=1000
     )
 
     await assert_config_state(ops_test, istio_core, policy_name)
@@ -172,10 +152,10 @@ async def test_istio_ingress_config_relation_break_and_recovery(ops_test: OpsTes
     # Re-establish the relation and verify the config state.
     await ops_test.model.add_relation(INGRESS_CONFIG_RELATION, APP_NAME)
     await ops_test.model.wait_for_idle(
-        apps=[APP_NAME, OAUTH2_K8S.application_name], status="active", timeout=1000
+        apps=[APP_NAME, oauth_k8s.application_name], status="active", timeout=1000
     )
     await ops_test.models.get(CORE_ISTIO_MODEL).model.wait_for_idle(
-        [ISTIO_K8S.application_name], status="active", timeout=1000
+        [istio_k8s.application_name], status="active", timeout=1000
     )
 
     await assert_config_state(ops_test, istio_core, policy_name)
@@ -225,7 +205,7 @@ async def assert_config_state(ops_test: OpsTest, istio_core, policy_name: str) -
     mesh_config = await load_mesh_config(istio_core.model.name)
     envoy_authz = get_envoy_authz(mesh_config, provider_name)
 
-    expected_service = f"{OAUTH2_K8S.application_name}.{ops_test.model.name}.svc.cluster.local"
+    expected_service = f"{oauth_k8s.application_name}.{ops_test.model.name}.svc.cluster.local"
     assert (
         envoy_authz.get("service") == expected_service
     ), f"Expected service '{expected_service}', got '{envoy_authz.get('service')}'"
